Move tests from the `rpc_types_gen` module into their appropriate module
